requirements.txt
alembic==1.10.2
aniso8601==9.0.1
attrs==23.1.0
Babel==2.12.1
blinker==1.4
cffi==1.16.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dnspython==2.3.0
email-validator==2.0.0.post1
emoji==1.7.0
filetype==1.2.0
Flask==2.2.3
Flask-Admin==1.6.1
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
flask-restx==1.1.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
gevent==23.9.0.post1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.1.7
mysql-connector-python==9.0.0
netaddr==0.8.0
OctoPrint-FileCheck==2021.2.23
OctoPrint-FirmwareCheck==2021.10.11
OctoPrint-PiSupport==2022.6.13
ordered-set==4.1.0
packaging==23.1
passlib==1.7.2
pathvalidate==2.5.2
peewee==3.13.1
Pillow==9.5.0
pkginfo==1.9.6
psutil==5.9.5
psycopg2==2.9.9
pycparser==2.21
Pygments==2.15.0
pylru==1.2.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3
sarge==0.1.6
semantic-version==2.10.0
setuptools==72.1.0
sgmllib3k==1.0.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.9
tornado==6.2
typing_extensions==4.12.2
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.15
watchdog==1.0.2
webassets==2.0
Werkzeug==2.2.3
wrapt==1.14.1
wtf-peewee==3.0.0
WTForms==3.0.1
zeroconf==0.33.4
zope.event==5.0
zope.interface==7.0.1
